### 2017-07-16 #315 @pjotrp
Hi @simonvh,

Thank your for your submission and thank you for the interesting software.

Before I start looking for a reviewer can you fill out the paper a bit more?
https://github.com/simonvh/genomepy/blob/master/paper/paper.md

It should be descriptive of the software and give a bit of context. Citations should go in a bib file similar to
http://joss.theoj.org/papers/10.21105/joss.00025. The sources for this are here:
https://github.com/genenetwork/genenetwork2/tree/master/doc/joss/2016

Giving it a nice presentation will help popularize your project and get citations.
