### 2017-07-18 #315 @pjotrp
OK, looking good. @peterjc can you suggest a biopython contributor as a reviewer for genomepy?
