### 2017-07-18 #315 @peterjc
He may decline as the author of a "rival" tool https://github.com/kblin/ncbi-genome-download (or the editors might consider this a conflict of interest), but I think that knowledge would let Kai Blin give a very well informed review.
