### 2017-07-18 #315 @pjotrp
Thanks @peterjc! @kblin can you take a look at this?
