### 2017-07-18 #315 @kblin
What's the timeline on this? I have a number of other pending reviews, and it looks like reviewing this needs a decent internet connection, so not something I can do while traveling.

At a first glance, I don't think I have a big conflict of interest, the focus of genomepy really seems to be more focused on interacting with the UCSC genomes and higher eukaryotes, whereas ncbi-genome-download is pretty much about what it says on the label, with a feature set more focused on the metadata available from NCBI.
