### 2017-07-18 #315 @simonvh
From my side at least there is no big rush. I'm leaving for a few weeks of holiday this Friday (pretty much off the grid). That means I won't have much time for a response before half of August anyway. In addition, I do agree, I also don't see any real conflict of interest.
