### 2017-07-19 #315 @pjotrp
Thanks @kblin. @simonvh, to expedite the review process can you check whether your submission matches our check list? See https://github.com/openjournals/joss-reviews/issues/307 for an example. That would be helpful, thanks!
