### 2017-07-19 #315 @simonvh
Thanks @pjotrp, I went through the checklist and added a few missing bits of information. I think it should be OK, looking forward to @kblin's input.
