### 2017-07-20 #320 @whedon
[REVIEW]: genomepy: download genomes the easy way
**Submitting author:** @simonvh (<a href="http://orcid.org/0000-0002-0411-3219">SJ van Heeringen</a>)
**Repository:** <a href="http://github.com/simonvh/genomepy" target ="_blank">http://github.com/simonvh/genomepy</a>
**Version:** 0.3.1
