### 2017-08-01 #320 @kblin
Looks good to me. Apart from the example in the README, I did not see documentation on using the API functions, but this is mainly a command line tool, and the CLI is documented nicely.
