### 2017-08-02 7eea7ce
Fixing JOSS paper
