### 2017-08-02 #320 @arfon
> @arfon we are ready to accept this submission.

Great! @simonvh could you please merge this PR https://github.com/simonvh/genomepy/pull/18 which fixes up the paper metadata.

Then, could you make an archive of the reviewed software in Zenodo/figshare/other service and update this thread with the DOI of the archive? I can then move forward with accepting the submission.
