### 2017-08-15 #320 @simonvh
Sorry @arfon, holidays without computer so it took a while. PR is now merged, thanks! The Zenodo DOI of this version is [10.5281/zenodo.831969](https://doi.org/10.5281/zenodo.831969).
